pipeline/requirements.txt
apache-airflow>=2.5.0
apache-airflow-providers-postgres
pandas>=1.5.0
numpy>=1.23.0
psycopg2-binary
